requirements.txt
# Dépendances principales
streamlit>=1.22.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
plotly>=5.14.0
pyyaml>=6.0
requests>=2.28.0
python-dotenv>=1.0.0

# API et intégrations
google-api-python-client>=2.80.0
google-auth>=2.16.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-cloud-storage>=2.8.0
pexels-api>=1.0.1

# Traitement audio et vidéo
moviepy>=1.0.3
ffmpeg-python>=0.2.0
pydub>=0.25.1
librosa>=0.10.0

# Synthèse vocale
torch>=2.0.0
torchaudio>=2.0.0
# Kokoro TTS sera installé séparément

# Génération d'images
diffusers>=0.14.0
transformers>=4.28.0
accelerate>=0.18.0

# Utilitaires
tqdm>=4.65.0
pillow>=9.5.0
beautifulsoup4>=4.12.0
nltk>=3.8.1
scikit-learn>=1.2.2
